fix model loading for transformers 5.x: dtype not torch_dtype
Confirmed against the actual installed stack on the cluster (GPU
smoke test: tokenizer, model load, generate all work with dtype=).
torch_dtype isn't just deprecated in transformers 5 — it's gone from
PreTrainedModel.from_pretrained's signature entirely. Pinned
requirements.txt to the versions actually running there, and noted
the --only-binary install quirks (pyarrow needs rust otherwise,
antlr4-python3-runtime has no wheel at all but builds fine).

+++ src/p2p_safety/model_utils.py
@@ -31,7 +31,11 @@ def load_model_and_tokenizer(model_name: str, dtype: str, device: str) -> tuple[
     # at the same position and `generate` doesn't need per-row trimming.
     tokenizer.padding_side = "left"
 
-    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch_dtype)
+    # transformers >=5 renamed from_pretrained's dtype kwarg from
+    # torch_dtype to dtype (confirmed against the installed version —
+    # torch_dtype is no longer in PreTrainedModel.from_pretrained's
+    # signature at all, not just deprecated).
+    model = AutoModelForCausalLM.from_pretrained(model_name, dtype=torch_dtype)
     model.to(device)
     model.eval()
     return model, tokenizer
